requirements.txt
llama-cpp-python
crewai